+++ testing/cffi0/backend_tests.py
@@ -121,6 +121,16 @@ class BackendTests:
         with pytest.raises(IndexError):
             p[-1] = 44
 
+    def test_array_slicing(self):
+        ffi = FFI(backend=self.Backend())
+        p = ffi.new("int[10]")
+        for i in range(10):
+            p[i] = i
+        assert repr(p[0:3]) == "<cdata 'int[]' sliced length 3>"
+        assert list(p[0:3]) == [0, 1, 2]
+        with pytest.raises(IndexError):
+            _ = p[0:11]
+
     def test_new_array_args(self):
         ffi = FFI(backend=self.Backend())
         # this tries to be closer to C: where we say "int x[5] = {10, 20, ..}"
@@ -1417,6 +1427,7 @@ class BackendTests:
         assert ffi.getctype("pe") == 'e *'
         assert ffi.getctype("e1*") == 'e1 *'
 
+    @pytest.mark.thread_unsafe
     def test_opaque_enum(self):
         import warnings
         ffi = FFI(backend=self.Backend())