+++ testing/cffi0/test_verify2.py
@@ -1,8 +1,11 @@
 import pytest
 from .test_verify import *
 
-# eliminate warning noise from common test modules that are repeatedly re-imported
-pytestmark = pytest.mark.filterwarnings("ignore:reimporting:UserWarning")
+pytestmark = [
+    pytest.mark.thread_unsafe,
+    # eliminate warning noise from common test modules that are repeatedly re-imported
+    pytest.mark.filterwarnings("ignore:reimporting:UserWarning"),
+]
 
 # This test file runs normally after test_verify.  We only clean up the .c
 # sources, to check that it also works when we have only the .so.  The